backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.9.1
python-dotenv==1.0.1
pydantic==2.5.0
openai>=1.13.0
tiktoken==0.5.2





